pyproject.toml
[project]
name = "planning-center-mcp"
version = "0.1.0"
description = "Planning Center MCP Server - A FastMCP server that provides tools for the Planning Center People API"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.10.6",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dateutil>=2.9.0",
    "nest-asyncio>=1.6.0",
]

[tool.uv]
dev-dependencies = []
